whisperlive-backend-main/requirements.txt
distlib==0.3.9
setuptools==78.1.0
# Whisper + Transcription
faster-whisper==0.10.0
numpy==1.26.4

# WebSocket & Real-time
python-socketio==5.11.1
websocket-client==1.7.0
eventlet==0.35.1

# Server
werkzeug==3.0.1
